pyproject.toml
[tool.poetry]
name = "gp-schedule-api"
version = "0.1.0"
description = "Web API"
authors = ["kartemuz <kartemuz@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10.13"
loguru = "^0.7.2"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
pytest = "^8.3.2"
email-validator = "^2.2.0"
aiomysql = "^0.2.0"
pytest-dotenv = "^0.5.2"
pydantic = "^2.8.2"
alembic = "^1.13.2"
pytest-async-sqlalchemy = "^0.2.0"
pytest-asyncio = "^0.23.8"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
